Enhance logging and reporting features with customizable time formats
Updated the logging module to support customizable time formats for webhook messages and hourly reports. Added new settings for 12-hour and 24-hour formats in the configuration, allowing users to choose their preferred display style. Refactored related classes and methods to accommodate these changes, ensuring consistent time formatting across the application. Additionally, introduced a new profiles tab in the web interface for better profile management.
