+++ src/modules/logging/webhook.py
@@ -5,7 +5,7 @@ from requests.exceptions import SSLError, ConnectionError
 last_message_id = None
 last_channel_id = None
 
-def webhook(url, title, desc, time, color, imagePath = None, ping_user_id = None):
+def webhook(url, title, desc, time, color, imagePath = None, ping_user_id = None, time_format=24):
     global last_message_id, last_channel_id
     webhook = DiscordWebhook(url = url,rate_limit_retry=True)
     
@@ -13,10 +13,22 @@ def webhook(url, title, desc, time, color, imagePath = None, ping_user_id = None
     if ping_user_id:
         webhook.content = f"<@{ping_user_id}>"
     
-    if title:
-        embed = DiscordEmbed(title="[{}] {}".format(time,title), description=desc, color=color)
+    # Format time based on preference
+    if time_format == 12:
+        try:
+            # Parse the time string and convert to 12-hour format
+            from datetime import datetime
+            time_obj = datetime.strptime(time, "%H:%M:%S")
+            formatted_time = time_obj.strftime("%I:%M:%S %p")
+        except:
+            formatted_time = time  # fallback to original if parsing fails
     else:
-        embed = DiscordEmbed(title="", description="[{}] {}".format(time,desc), color=color)
+        formatted_time = time
+
+    if title:
+        embed = DiscordEmbed(title="[{}] {}".format(formatted_time,title), description=desc, color=color)
+    else:
+        embed = DiscordEmbed(title="", description="[{}] {}".format(formatted_time,desc), color=color)
     #if to add image
     if imagePath:
         with open(imagePath, "rb") as f:
