+++ src/webapp/htmlImports/persistent/tabs.html
@@ -1,42 +1,87 @@
-<ul style="width: 5.3rem; background-color: #202225; display: flex; flex-direction: column; align-items: center; padding-top: 1rem; gap: 3%; margin-top: 0.1rem;">
-    <li class = "sidebar-item active" id = "home-tab" onclick="loadTasks()">
-        <img src="./assets/sidebar/home.png">
-        Home
-    </li>
+<ul
+  style="
+    width: 5.3rem;
+    background-color: #202225;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding-top: 1rem;
+    gap: 3%;
+    margin-top: 0.1rem;
+  "
+>
+  <li class="sidebar-item active" id="home-tab" onclick="loadTasks()">
+    <img src="./assets/sidebar/home.png" />
+    Home
+  </li>
 
-    <li class = "sidebar-item" id = "gather-tab">
-        <img src="./assets/sidebar/gather.png">
-        Gather
-    </li>
+  <li class="sidebar-item" id="gather-tab">
+    <img src="./assets/sidebar/gather.png" />
+    Gather
+  </li>
 
-    <li class = "sidebar-item" id = "collect-tab">
-        <img src="./assets/sidebar/collect.png">
-        Collect
-    </li>
+  <li class="sidebar-item" id="collect-tab">
+    <img src="./assets/sidebar/collect.png" />
+    Collect
+  </li>
 
-    <li class = "sidebar-item" id = "kill-tab">
-        <img src="./assets/sidebar/kill.png">
-        Kill
-    </li>
+  <li class="sidebar-item" id="kill-tab">
+    <img src="./assets/sidebar/kill.png" />
+    Kill
+  </li>
 
-    <li class = "sidebar-item" id = "boost-tab">
-        <img src="./assets/sidebar/boost.png">
-        Boost
-    </li>
+  <li class="sidebar-item" id="boost-tab">
+    <img src="./assets/sidebar/boost.png" />
+    Boost
+  </li>
 
-    <li class = "sidebar-item" id = "planters-tab">
-        <img src="./assets/sidebar/planters.png">
-        Planters
-    </li>
+  <li class="sidebar-item" id="planters-tab">
+    <img src="./assets/sidebar/planters.png" />
+    Planters
+  </li>
 
-    <li class = "sidebar-item" id = "quests-tab">
-        <img src="./assets/sidebar/quests.png">
-        Quests
-    </li>
+  <li class="sidebar-item" id="quests-tab">
+    <img src="./assets/sidebar/quests.png" />
+    Quests
+  </li>
 
-    <!--https://icon-sets.iconify.design/grommet-icons/configure/-->
-    <li class = "sidebar-item" id = "config-tab">
-        <img src="./assets/sidebar/config.png">
-        Config
-    </li>
+  <!--https://icon-sets.iconify.design/grommet-icons/configure/-->
+  <li class="sidebar-item" id="config-tab">
+    <img src="./assets/sidebar/config.png" />
+    Config
+  </li>
+
+  <li class="sidebar-item" id="profiles-tab">
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      style="width: 1.4rem; height: 1.4rem"
+    >
+      <path
+        d="M12 2L2 7L12 12L22 7L12 2Z"
+        stroke="white"
+        stroke-width="2"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+      <path
+        d="M2 17L12 22L22 17"
+        stroke="white"
+        stroke-width="2"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+      <path
+        d="M2 12L12 17L22 12"
+        stroke="white"
+        stroke-width="2"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+    </svg>
+    Profiles
+  </li>
 </ul>
