+++ src/webapp/htmlImports/tabs/profiles.html
@@ -1,0 +1,240 @@
+<div style="display: flex; justify-content: center; width: 100%">
+  <div
+    style="width: 74%; overflow-y: auto; height: 94vh; display: block"
+    id="profiles-content"
+  >
+    <style>
+      .profile-card {
+        background: #36393f;
+        border-radius: 8px;
+        padding: 1rem;
+        margin-bottom: 1rem;
+        border: 1px solid #4f545c;
+      }
+      .profile-card-header {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        margin-bottom: 0.5rem;
+      }
+      .profile-card-title {
+        color: #dcddde;
+        font-size: 1.1rem;
+        font-weight: 600;
+      }
+      .profile-list {
+        display: flex;
+        flex-direction: column;
+        gap: 0.5rem;
+        max-height: 300px;
+        overflow-y: auto;
+      }
+      .profile-item {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        background: #2f3136;
+        padding: 0.75rem 1rem;
+        border-radius: 6px;
+        border: 1px solid transparent;
+        transition: all 0.2s ease;
+      }
+      .profile-item:hover {
+        border-color: #7a77bb;
+      }
+      .profile-item.active {
+        border-color: #77bb7a;
+        background: #2a3a2a;
+      }
+      .profile-item-name {
+        color: #dcddde;
+        font-weight: 500;
+      }
+      .profile-item-actions {
+        display: flex;
+        gap: 0.5rem;
+      }
+      .profile-btn {
+        padding: 0.4rem 0.8rem;
+        border-radius: 4px;
+        border: none;
+        cursor: pointer;
+        font-size: 0.85rem;
+        font-family: inherit;
+        transition: all 0.2s ease;
+      }
+      .profile-btn-primary {
+        background: #7a77bb;
+        color: white;
+      }
+      .profile-btn-primary:hover {
+        background: #5c59a3;
+      }
+      .profile-btn-success {
+        background: #77bb7a;
+        color: white;
+      }
+      .profile-btn-success:hover {
+        background: #5a9a5c;
+      }
+      .profile-btn-danger {
+        background: #f04747;
+        color: white;
+      }
+      .profile-btn-danger:hover {
+        background: #d84040;
+      }
+      .profile-btn-secondary {
+        background: #4f545c;
+        color: #dcddde;
+      }
+      .profile-btn-secondary:hover {
+        background: #686d73;
+      }
+      .profile-input {
+        background: #2f3136;
+        border: 1px solid #4f545c;
+        border-radius: 4px;
+        padding: 0.5rem 0.75rem;
+        color: #dcddde;
+        font-size: 0.95rem;
+        width: 200px;
+        font-family: inherit;
+      }
+      .profile-input:focus {
+        outline: none;
+        border-color: #7a77bb;
+      }
+      .profile-status {
+        padding: 0.5rem 1rem;
+        border-radius: 4px;
+        margin-top: 0.5rem;
+        font-size: 0.9rem;
+      }
+      .profile-status.success {
+        background: rgba(119, 187, 122, 0.2);
+        color: #77bb7a;
+        border: 1px solid #77bb7a;
+      }
+      .profile-status.error {
+        background: rgba(240, 71, 71, 0.2);
+        color: #f04747;
+        border: 1px solid #f04747;
+      }
+      .current-profile-badge {
+        background: #77bb7a;
+        color: white;
+        padding: 0.2rem 0.5rem;
+        border-radius: 4px;
+        font-size: 0.75rem;
+        font-weight: 600;
+        margin-left: 0.5rem;
+      }
+    </style>
+    <div
+      class="poppins-medium standard-container"
+      style="display: block; justify-items: unset; padding-top: 1rem"
+    >
+      <h2>Profile Management</h2>
+      <p style="font-weight: 500; font-size: 1rem">
+        Save and switch between different configuration profiles. Each profile
+        stores your settings, field configurations, and task priorities
+        separately.
+      </p>
+      <div class="seperator"></div>
+
+      <!-- Current Profile Section -->
+      <div class="profile-card">
+        <div class="profile-card-header">
+          <span class="profile-card-title">📁 Current Profile</span>
+        </div>
+        <div style="display: flex; align-items: center; gap: 1rem">
+          <span
+            id="current-profile-display"
+            style="color: #77bb7a; font-size: 1.2rem; font-weight: 600"
+            >Loading...</span
+          >
+        </div>
+      </div>
+
+      <!-- Switch Profile Section -->
+      <div class="profile-card">
+        <div class="profile-card-header">
+          <span class="profile-card-title">🔄 Switch Profile</span>
+        </div>
+        <p style="color: #b9bbbe; margin-bottom: 1rem; font-size: 0.9rem">
+          Select a profile to switch to. Your current settings will be saved
+          automatically.
+        </p>
+        <div class="profile-list" id="profile-list">
+          <!-- Profile items will be loaded dynamically -->
+        </div>
+      </div>
+
+      <!-- Create Profile Section -->
+      <div class="profile-card">
+        <div class="profile-card-header">
+          <span class="profile-card-title">➕ Create New Profile</span>
+        </div>
+        <p style="color: #b9bbbe; margin-bottom: 1rem; font-size: 0.9rem">
+          Create a new profile based on your current settings.
+        </p>
+        <div style="display: flex; gap: 0.75rem; align-items: center">
+          <input
+            type="text"
+            id="new-profile-name"
+            class="profile-input"
+            placeholder="Enter profile name..."
+          />
+          <button
+            class="profile-btn profile-btn-success"
+            onclick="createNewProfile()"
+          >
+            Create Profile
+          </button>
+        </div>
+        <div id="create-profile-status"></div>
+      </div>
+
+      <!-- Duplicate Profile Section -->
+      <div class="profile-card">
+        <div class="profile-card-header">
+          <span class="profile-card-title">📋 Duplicate Profile</span>
+        </div>
+        <p style="color: #b9bbbe; margin-bottom: 1rem; font-size: 0.9rem">
+          Create a copy of an existing profile with a new name.
+        </p>
+        <div
+          style="
+            display: flex;
+            gap: 0.75rem;
+            align-items: center;
+            flex-wrap: wrap;
+          "
+        >
+          <select
+            id="duplicate-source-profile"
+            class="profile-input"
+            style="width: 150px"
+          >
+            <!-- Options will be loaded dynamically -->
+          </select>
+          <span style="color: #b9bbbe">→</span>
+          <input
+            type="text"
+            id="duplicate-new-name"
+            class="profile-input"
+            placeholder="New profile name..."
+          />
+          <button
+            class="profile-btn profile-btn-primary"
+            onclick="duplicateExistingProfile()"
+          >
+            Duplicate
+          </button>
+        </div>
+        <div id="duplicate-profile-status"></div>
+      </div>
+    </div>
+  </div>
+</div>
