+++ src/webapp/index.html
@@ -1,27 +1,74 @@
 <!DOCTYPE html>
 <html>
-<head>
+  <head>
     <title>exihs Macro</title>
-    <link rel="stylesheet" href="styles/style.css">
-    <link rel="shortcut icon" type="image/x-icon" href="assets/general/icon.png" />
+    <link rel="stylesheet" href="styles/style.css" />
+    <link
+      rel="shortcut icon"
+      type="image/x-icon"
+      href="assets/general/icon.png"
+    />
     <script type="text/javascript" src="./scripts/jquery.js"></script>
     <script type="text/javascript" src="./scripts/ui-builder.js"></script>
     <script type="text/javascript" src="./scripts/utilities.js"></script>
     <script type="text/javascript" src="./scripts/emojis.js"></script>
-</head>
-<body>
-    <div id = "header-placeholder"></div>
+  </head>
+  <body>
+    <div id="header-placeholder"></div>
     <main>
-        <div id = "tabs-placeholder"></div>
-        <div id="home-placeholder" class="content" style="display: flex; padding-top: 0.5rem; padding-bottom: 0.8rem; width: 100%; margin-left: 1rem; margin-right: 1rem;"></div>
-        <div id="gather-placeholder" class="content" style="width: 100%; display: none;"></div>
-        <div id="collect-placeholder" class="content" style="width: 100%; display: none;"></div>
-        <div id="config-placeholder" class="content" style="width: 100%; display: none;"></div>
-        <div id="kill-placeholder" class="content" style="width: 100%; display: none;"></div>
-        <div id="planters-placeholder" class="content" style="width: 100%; display: none; flex-direction: column;"></div>
-        <div id="boost-placeholder" class="content" style="width: 100%; display: none;"></div>
-        <div id="quests-placeholder" class="content" style="width: 100%; display: none;"></div>
-
+      <div id="tabs-placeholder"></div>
+      <div
+        id="home-placeholder"
+        class="content"
+        style="
+          display: flex;
+          padding-top: 0.5rem;
+          padding-bottom: 0.8rem;
+          width: 100%;
+          margin-left: 1rem;
+          margin-right: 1rem;
+        "
+      ></div>
+      <div
+        id="gather-placeholder"
+        class="content"
+        style="width: 100%; display: none"
+      ></div>
+      <div
+        id="collect-placeholder"
+        class="content"
+        style="width: 100%; display: none"
+      ></div>
+      <div
+        id="config-placeholder"
+        class="content"
+        style="width: 100%; display: none"
+      ></div>
+      <div
+        id="kill-placeholder"
+        class="content"
+        style="width: 100%; display: none"
+      ></div>
+      <div
+        id="planters-placeholder"
+        class="content"
+        style="width: 100%; display: none; flex-direction: column"
+      ></div>
+      <div
+        id="boost-placeholder"
+        class="content"
+        style="width: 100%; display: none"
+      ></div>
+      <div
+        id="quests-placeholder"
+        class="content"
+        style="width: 100%; display: none"
+      ></div>
+      <div
+        id="profiles-placeholder"
+        class="content"
+        style="width: 100%; display: none"
+      ></div>
     </main>
     <script type="text/javascript" src="/eel.js"></script>
     <script src="./scripts/script.js"></script>
@@ -34,5 +81,6 @@
     <script src="./scripts/tabs/planters.js"></script>
     <script src="./scripts/tabs/boost.js"></script>
     <script src="./scripts/tabs/quests.js"></script>
-</body>
+    <script src="./scripts/tabs/profiles.js"></script>
+  </body>
 </html>
